backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
eventlet==0.33.3
python-socketio==5.8.0
PyJWT
